+++ dca/tl_anystores_settings.php
@@ -25,7 +25,7 @@ $GLOBALS['TL_DCA']['tl_anystores_settings'] = array
     (
         'default' =>
         '
-            {anystores_common_legend},anystores_defaultMarker;
+            {anystores_common_legend},anystores_defaultMarker,anystores_tableHeaders;
             {anystores_api_legend},anystores_geoApi,anystores_apiKey,anystores_apiBrowserKey
         '
     ),
@@ -83,7 +83,39 @@ $GLOBALS['TL_DCA']['tl_anystores_settings'] = array
                 'helpwizard' => true,
                 'tl_class'   => 'w50'
             )
+        ),
+        'anystores_tableHeaders' => array
+        (
+            'label'            => &$GLOBALS['TL_LANG']['tl_anystores_settings']['anystores_tableHeaders'],
+            'inputType'        => 'checkboxWizard',
+            'options_callback' => array('tl_anystores_settings', 'getTableHeader'),
+            'eval'             => array
+            (
+                'multiple' => true,
+            )
         )
     )
 
 );
+
+
+class tl_anystores_settings
+{
+
+    public function getTableHeader(\DataContainer $dc)
+    {
+        \System::loadLanguageFile('tl_anystores');
+
+        $arrLabels = array();
+
+        $arrFields = \DcaExtractor::getInstance('tl_anystores')->getFields();
+
+        foreach($arrFields as $key => $value)
+        {
+            $arrLabels[$key] = $GLOBALS['TL_LANG']['tl_anystores'][$key][0] ?: $key;
+        };
+
+        return $arrLabels;
+    }
+
+}